clean and move to description
